Backend - Adjustments to how repositionEntity() works.
Added support for -1 value to positionEntity() where it'll place such Entities at the end (so we don't need to know the length of the array we're appending to).
Solved big problem with off-by-one errors
Renaming repositionEntity() to just positionEntity().
Updated and improved unit tests for positionEntity().
Renamed interface Sortable to SortableEntity.
Defined a new type for lexical_order named LexicalOrder. It's just for readability.
Removed some code I meant to remove before.