Frontend - Mutations to move and delete Categories. Lots of edits to way I manage mutations and their optimistic updates.
This is basically MVP++. The additional feature being the ability to create, move, and delete Categories.

What I still want to do is, in this order:
Support showing created/updated times on a click.
Support editing Task title in the front end.
Support editing Category title in the front end.
Support description for Task in the front and back end.
Support drag-and-drop.
Add support for Tags in the front and back end.